GetNodesFromIDs and GetNodesEdges methods for Graph
Added two new methods for Graph and Backend interface.
These methods are useful to reduce the number of queries while using
backends with history, improving speed and reducing load in the backend.

GetNodesEdges, return the list with all edges for a list of nodes
GetNodesFromIDs, return the list of nodes given a list of identifiers.

Batching is used to avoid hitting the max number of clauses set by ES
(is set to the default value of 512).