Fix event handler filling channel and being blocked
As explained in the previous commit, the goroutine consuming g.eventChan
channel could get blocked if a linker generate more than maxEvents
events.
Fix this problem discarding events when the channel is full.
Return an error if this point is reached.

Increase the number of maxEvents to reduce the probability of event
discard.