ES: fix handling of aliases while starting
Skydive configured with ElasticSearch was deleting correct indices from
the alias. This was also causing an error while trying to create the
next archive index.

The error was considering all indices in the alias as incorrect.
Now, only indices with an old schema version are considered old.
If there is currently an index ponting to the alias, that one should be
considered as the valid one.

Fixes #2404