Graffiti: subgraph edge filtering
When using Subgraph() with nodes (G.V().SubGraph()), now it is possible
to filter which edges we want to return, instead of all connecting the
selected nodes.

For example, if we have two nodes, A and B, and two edges between them,
"e1" and "e1", now we can use this syntaxis to just return the edge
"e1".

G.V().SubGraph("Name", "e1")

The syntaxis inside of SubGraph is the same as in 'Has'.